requirements.txt
fastapi
sqlmodel
psycopg2-binary
python-dotenv
uvicorn